% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_load.R
\name{m_load}
\alias{m_load}
\title{Carregar pacotes}
\usage{
m_load(pkg)
}
\arguments{
\item{pkg}{o nome de um pacote ou um vetor de string com o nome de varios pacotes.}
}
\value{
Pacotes anexados
}
\description{
Carregar, instalar e atualizar pacotes
}
\details{
A funcao verifica se o pacote esta instalado e atualizado. Caso negativo, instalara a versao mais atualizada do pacote.
Em seguida, carrega todos os pacotes nomeados.
}
\examples{

Caregando apenas um pacote

m_load("beepr")

# Vetor contento varios pacotes 

 pkg <- c("beepr", "magrittr")
 m_load(pkg)

# Ocultar mensagens desnecessarias 

pkg <- c("beepr", "magrittr")
invisible(m_load(pkg))

}
\author{
Mourao
}
