% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_clear.R, R/m_load.R
\name{m_load}
\alias{m_load}
\alias{m_clear}
\title{Limpar ambiente}
\usage{
m_clear(manter = NULL, console = TRUE)

m_load(pkg)
}
\arguments{
\item{manter}{Lista contendo o nome das variaveis, funcoes e objetos a serem mantidos}

\item{console}{Limpar o console (TRUE) ou nao (FALSE). Funciona apenas no RStudio do Windows, assim
deve-se usar console == FALSE em outros sistemas.}

\item{pkg}{o nome de um pacote ou um vetor de string com o nome de varios pacotes.}
}
\value{
ambiente limpo

Pacotes anexados
}
\description{
Limpa o ambiente global (Global Environment)

Carregar, instalar e atualizar pacotes
}
\details{
A funcao retorna o ambiente mais limpo, podendo assim evitar erros no processamento dos dados.

A funcao verifica se o pacote esta instalado e atualizado. Caso negativo, instalara a versao mais atualizada do pacote.
Em seguida, carrega todos os pacotes nomeados.
}
\examples{

# Exemplo 1 

a <- 1
b <- "teste"
c <-  c(1,2,3)
d <- 456L

reset(manter = c("a", "b"))

# Exemplo 2

a <- 1
b <- "teste"
c <-  c(1,2,3)
d <- 456L

reset(manter = c("a", "b"), console = FALSE)


Caregando apenas um pacote

m_load("beepr")

# Vetor contento varios pacotes 

 pkg <- c("beepr", "magrittr")
 m_load(pkg)

# Ocultar mensagens desnecessarias 

pkg <- c("beepr", "magrittr")
invisible(m_load(pkg))

}
\author{
Mourao

Mourao
}
